+++ tests/calInfoTest.py
@@ -4,8 +4,8 @@
 ander906@purdue.edu
 HybridMEMS
 """
-
-import AgilentPNAXUtils as pnaUtils
+from context import pymeasrf
+import pymeasrf.AgilentPNAXUtils as pnaUtils
 
 pna = pnaUtils.AgilentPNAx('TCPIP0::192.168.1.1::inst0::INSTR')
 pna.getCalInfo()
